skip_bytes use File Seek
Currently for bytes skipping, tsMuxer reads 2 MB blocks at a time from the input file through a loop: this can take substantial time for large skips/files -especially when the file is on a NAS, accessed via WiFi.

This pull introduces file seeking so that only two blocks  (4MB) are left to be read.

Fixes issue #733.